feat(rmcp): separate wire from rest of implementation for improved separation of concerns.
