+++ src/app/mod.rs
@@ -1,4 +1,9 @@
 mod get_device_id;
 pub use get_device_id::{DeviceId, GetDeviceId};
 
+mod get_channel_authentication_capabilities;
+pub use get_channel_authentication_capabilities::{
+    Channel, ChannelAuthenticationCapabilities, GetChannelAuthenticationCapabilities,
+};
+
 pub mod auth;
